test claim role

+++ helper/claimRole.js
@@ -9,9 +9,11 @@ const web3 = new Web3(
     new Web3.providers.WebsocketProvider(rpcURL)
 );
 
-const { guildId } = require('../config.json');
 
-const assignRole = (user, amountStaked, wallet) => {
+const { guildId } = require('../config.json');
+const { data } = require('../commands/character.js');
+
+const assignRole = (user, amountStaked, wallet,client) => {
 
     if (amountStaked >= 100 && amountStaked < 500) {
         user.roles.add('880394877202997298'), user.roles.remove('883601505033269309')
@@ -29,7 +31,7 @@ const assignRole = (user, amountStaked, wallet) => {
     channel.send("<@" + user + "> Your Role has been successfully assigned");
 }
 
-async function ClaimRole(wallet, mes, userID) {
+async function ClaimRole(wallet, mes, userID,client) {
     const Address = web3.eth.accounts.recover("ForLootAndGlory Claim Role!", mes);
     const address = Address.toString().toLowerCase();
     console.log('recover address:', address)
@@ -38,7 +40,7 @@ async function ClaimRole(wallet, mes, userID) {
         let balance = _balance * 10 ** 18;
         const guild = await client.guilds.fetch(guildId)
         const user = await guild.members.fetch(userID)
-        assignRole(user, balance, wallet);
+        assignRole(user, balance, wallet,client);
         console.log('success recover address')
     } else {
         console.log('failed to recover address')
@@ -52,9 +54,27 @@ async function updateWhitelist(user, balance, wallet) {
             wallet: wallet,
             balance: balance
         };
-        fs.writeFile(`../whitelist/${user}.json`, obj, (err) => {
-            console.log(err)
-        })
+        fs.readdir(dir, (err, files) => {
+            console.log(files.length);
+            len = files.length;
+        });
+        let isExist = false;
+        for (let i = 0; i < files.length; i++) {
+            fs.readFile(`../whitelist/${files[i]}`, function read(err, data) {
+                if (err) {
+                    throw err;
+                }
+            })
+            if (data.wallet === wallet) {
+                isExist = true
+            }
+
+        }
+        if (isExist === false) {
+            fs.writeFile(`../whitelist/${user}.json`, obj, (err) => {
+                console.log(err)
+            })
+        }
     }
 }
 
@@ -67,7 +87,7 @@ async function createWhitelistfile() {
         len = files.length;
     });
     for (let i = 0; i < files.length; i++) {
-       fs.readFile(`../whitelist/${files[i]}`, function read(err, data) {
+        fs.readFile(`../whitelist/${files[i]}`, function read(err, data) {
             if (err) {
                 throw err;
             }
@@ -85,5 +105,6 @@ async function createWhitelistfile() {
 }
 
 module.exports = {
-    ClaimRole
+    ClaimRole,
+    createWhitelistfile
 }
